Switch to hosted auth and run cleanup (#13)
* Switch to hosted auth and run cleanup

* Save working

* User updates to authentik

* Fix saving

* Hide emails

* Release 1.0.0

* [check-spelling] Update metadata

Update for https://github.com/SpecialThing44/recipe-website/actions/runs/20344912011/attempts/1
Accepted in https://github.com/SpecialThing44/recipe-website/pull/13#issuecomment-3671276428

on-behalf-of: @check-spelling <check-spelling-bot@check-spelling.dev>

* Release 1.0.0

---------

+++ cooking-domain/src/main/scala/domain/ingredients/Quantity.scala
@@ -5,7 +5,7 @@ import io.circe.{Decoder, Encoder}
 
 case class Quantity(
     unit: Unit,
-    amount: Int
+    amount: Double
 )
 
 object Quantity {
